requirements.txt
flask
requests
python-dotenv
redis
celery